data/ split, fix refactor, make it ready for drop-in repalcements
