+++ README.md
@@ -4,25 +4,97 @@ CX Object System
 [![Build Status](https://travis-ci.org/skycoin/cxo.svg)](https://travis-ci.org/skycoin/cxo)
 [![Coverage Status](https://coveralls.io/repos/skycoin/cxo/badge.svg?branch=master)](https://coveralls.io/r/skycoin/cxo?branch=master)
 [![GoReportCard](https://goreportcard.com/badge/skycoin/cxo)](https://goreportcard.com/report/skycoin/cxo)
+[![Telegram group link](telegram-group.svg)](https://t.me/joinchat/B_ax-A6oCR9eQuAPiJtvaw)
 
-The CXO is ...
+![logo](cxo_logo.png)
 
-## Packages
+The CXO is objects system, goal of which is sharing any objects. The CXO
+is low level and designed to build application on top of it
 
-- `cmd/cxod` - CXO daemon (example node)
-- `cmd/cxocli` - cli to control the CXO daemon
-- `data` - CX obejcts database
-- `node` - CXO networking core
-- `skyobejct` - CX objects tree
+### How it wroks
+
+So, there is an owner who want to publicate something. This man creates feed,
+and adds a Root (R1) obejct to this feed. The Root refers to another obejcts.
+It's tree of any objects. The tree is immutable and object are immutable too.
+But the owner can update the tree publishing new version of the Root (R2).
+This root is based on previous version. This R2 is replacement for the R1.
+Thus, if the owner publishes some `Root` and updates it R1->R2->R3->R4. Any
+node that receive this feed can ignore R1,R2 and R3, because they are outdated
+since R4 borns
+
+For example
+```go
+type User struct {
+    Name string
+    Age  uint32
+    Bio  skyobject.Ref `schema:"pkg.Bio"`
+}
+
+type Bio struct {
+    Height uint32
+    Weight uint32
+}
+```
+
+It's like
+```go
+type User struct {
+   Name string
+   Age  int
+   Bio  *Bio
+}
+```
+
+The Ref is SHA256 reference of encoded value. Thus, we are using uint32 instead
+of int. Because the int can be int32 or int64 and not acceptable for
+encoding and decoding. E.g. the `User` will point to some encoded `Bio`.
+
+And if this `Bio` has been changed, then the `User` have to be changed too.
+And the owner, who want to share this information, creates feed. Creates Root
+obejct. Attaches actual instance of `User` to the `Root`. Signs the `Root` and
+publishes it. And when the owner want to update this information, he publishes
+new version of this `Root`.
+
+Thus, the `Root` can point to many objects. But, updating one obejct, we
+updates only one branch of the objects tree. All other obejcts already
+replicated by peers.
+
+Trusting. The owner has secret key, and feed is public key based
+on the secret key. And every Root is signed by the owner. Thus, a node,
+receiving a Root, can verify singature and be sure that this Root comes from
+that owner. This way if we know a feed (public key) then we can trust all
+received data from its feed.
 
 ### Version
 
-Use [dep](https://github.com/golang/dep) to use particular commit
+Use [dep](https://github.com/golang/dep) to use particular commit.
+
+To get mainline use
+```
+go get -u -t github.com/skycoin/cxo/...
+```
+Test them all
+```
+go test -cover -race github.com/skycoin/cxo/...
+```
+
+### Get started. How it works? Documentation
+
+See [CXO wiki](../../wiki) to get this information
+
+### Help and contacts
+
+- [telegram group](https://t.me/joinchat/B_ax-A6oCR9eQuAPiJtvaw)
+
+
+
+
+<!--
 
 ## Database
 
-- [data/README.md](https://github.com/logrusorgru/cxo/blob/master/data/README.md)
-- Binary data representation: [data/ABI.md](https://github.com/logrusorgru/cxo/blob/master/data/ABI.md)
+- [data/README.md](data/README.md)
+- Binary data representation: [data/ABI.md](data/ABI.md)
 
 ## Skyobject
 
@@ -331,3 +403,6 @@ TODO (kostyarin): explain detailed
 ##### Double-Linked List Example
 
 TODO (kostyarin): do it
+
+
+-->
