+++ data/cxds/memory.go
@@ -4,9 +4,11 @@ import (
 	"sync"
 
 	"github.com/skycoin/skycoin/src/cipher"
+
+	"github.com/skycoin/cxo/data"
 )
 
-type MemoryCXDS struct {
+type memoryCXDS struct {
 	mx  sync.RWMutex
 	kvs map[cipher.SHA256]memoryObject
 }
@@ -17,15 +19,13 @@ type memoryObject struct {
 	val []byte
 }
 
-// NewMemoryCXDS creates databse in memory.
-// The database based on golang map. The
-// datbase implements CXDS interface of
-// github.com/skycoin/cxo/data package
-func NewMemoryCXDS() *MemoryCXDS {
-	return &MemoryCXDS{kvs: make(map[cipher.SHA256]memoryObject)}
+// NewMemoryCXDS creates CXDS-databse in
+// memory. The database based on golang map
+func NewMemoryCXDS() data.CXDS {
+	return &memoryCXDS{kvs: make(map[cipher.SHA256]memoryObject)}
 }
 
-func (m *MemoryCXDS) Get(key cipher.SHA256) (val []byte, rc uint32, err error) {
+func (m *memoryCXDS) Get(key cipher.SHA256) (val []byte, rc uint32, err error) {
 	m.mx.RLock()
 	defer m.mx.RUnlock()
 
@@ -33,11 +33,11 @@ func (m *MemoryCXDS) Get(key cipher.SHA256) (val []byte, rc uint32, err error) {
 		val, rc = mo.val, mo.rc
 		return
 	}
-	err = ErrNotFound
+	err = data.ErrNotFound
 	return
 }
 
-func (m *MemoryCXDS) GetInc(key cipher.SHA256) (val []byte, rc uint32,
+func (m *memoryCXDS) GetInc(key cipher.SHA256) (val []byte, rc uint32,
 	err error) {
 
 	m.mx.Lock()
@@ -50,11 +50,11 @@ func (m *MemoryCXDS) GetInc(key cipher.SHA256) (val []byte, rc uint32,
 		m.kvs[key] = mo // save
 		return
 	}
-	err = ErrNotFound
+	err = data.ErrNotFound
 	return
 }
 
-func (m *MemoryCXDS) Set(key cipher.SHA256, val []byte) (rc uint32, err error) {
+func (m *memoryCXDS) Set(key cipher.SHA256, val []byte) (rc uint32, err error) {
 	m.mx.Lock()
 	defer m.mx.Unlock()
 
@@ -73,7 +73,7 @@ func (m *MemoryCXDS) Set(key cipher.SHA256, val []byte) (rc uint32, err error) {
 	return
 }
 
-func (m *MemoryCXDS) Add(val []byte) (key cipher.SHA256, rc uint32, err error) {
+func (m *memoryCXDS) Add(val []byte) (key cipher.SHA256, rc uint32, err error) {
 	if len(val) == 0 {
 		err = ErrEmptyValue
 		return
@@ -83,7 +83,7 @@ func (m *MemoryCXDS) Add(val []byte) (key cipher.SHA256, rc uint32, err error) {
 	return
 }
 
-func (m *MemoryCXDS) Inc(key cipher.SHA256) (rc uint32, err error) {
+func (m *memoryCXDS) Inc(key cipher.SHA256) (rc uint32, err error) {
 	m.mx.Lock()
 	defer m.mx.Unlock()
 
@@ -93,11 +93,11 @@ func (m *MemoryCXDS) Inc(key cipher.SHA256) (rc uint32, err error) {
 		rc = mo.rc
 		return
 	}
-	err = ErrNotFound
+	err = data.ErrNotFound
 	return
 }
 
-func (m *MemoryCXDS) Dec(key cipher.SHA256) (rc uint32, err error) {
+func (m *memoryCXDS) Dec(key cipher.SHA256) (rc uint32, err error) {
 	m.mx.Lock()
 	defer m.mx.Unlock()
 
@@ -111,11 +111,11 @@ func (m *MemoryCXDS) Dec(key cipher.SHA256) (rc uint32, err error) {
 		m.kvs[key] = mo
 		return
 	}
-	err = ErrNotFound
+	err = data.ErrNotFound
 	return
 }
 
-func (m *MemoryCXDS) DecGet(key cipher.SHA256) (val []byte, rc uint32,
+func (m *memoryCXDS) DecGet(key cipher.SHA256) (val []byte, rc uint32,
 	err error) {
 
 	m.mx.Lock()
@@ -132,11 +132,11 @@ func (m *MemoryCXDS) DecGet(key cipher.SHA256) (val []byte, rc uint32,
 		m.kvs[key] = mo
 		return
 	}
-	err = ErrNotFound
+	err = data.ErrNotFound
 	return
 }
 
-func (m *MemoryCXDS) MultiGet(keys []cipher.SHA256) (vals [][]byte, err error) {
+func (m *memoryCXDS) MultiGet(keys []cipher.SHA256) (vals [][]byte, err error) {
 	if len(keys) == 0 {
 		return
 	}
@@ -150,13 +150,13 @@ func (m *MemoryCXDS) MultiGet(keys []cipher.SHA256) (vals [][]byte, err error) {
 			vals = append(vals, mo.val)
 			continue
 		}
-		err = ErrNotFound
+		err = data.ErrNotFound
 		return
 	}
 	return
 }
 
-func (m *MemoryCXDS) MultiAdd(vals [][]byte) (err error) {
+func (m *memoryCXDS) MultiAdd(vals [][]byte) (err error) {
 	for _, val := range vals {
 		if _, _, err = m.Add(val); err != nil {
 			return
@@ -165,7 +165,7 @@ func (m *MemoryCXDS) MultiAdd(vals [][]byte) (err error) {
 	return
 }
 
-func (m *MemoryCXDS) MultiInc(keys []cipher.SHA256) (err error) {
+func (m *memoryCXDS) MultiInc(keys []cipher.SHA256) (err error) {
 	for _, key := range keys {
 		if _, err = m.Inc(key); err != nil {
 			return
@@ -174,7 +174,7 @@ func (m *MemoryCXDS) MultiInc(keys []cipher.SHA256) (err error) {
 	return
 }
 
-func (m *MemoryCXDS) MultiDec(keys []cipher.SHA256) (err error) {
+func (m *memoryCXDS) MultiDec(keys []cipher.SHA256) (err error) {
 	for _, key := range keys {
 		if _, err = m.Dec(key); err != nil {
 			return
@@ -183,7 +183,25 @@ func (m *MemoryCXDS) MultiDec(keys []cipher.SHA256) (err error) {
 	return
 }
 
-func (m *MemoryCXDS) Close() (_ error) {
+func (m *memoryCXDS) Iterate(iterateFunc func(cipher.SHA256,
+	uint32) error) (err error) {
+
+	m.mx.RLock()
+	defer m.mx.RUnlock()
+
+	for k, mo := range m.kvs {
+		if err = iterateFunc(k, mo.rc); err != nil {
+			if err == data.ErrStopIteration {
+				err = nil
+			}
+			return
+		}
+	}
+
+	return
+}
+
+func (m *memoryCXDS) Close() (_ error) {
 	m.mx.Lock()
 	defer m.mx.Unlock()
