+++ data/data.go
@@ -1,61 +1,31 @@
 package data
 
 import (
-	"github.com/skycoin/skycoin/src/cipher"
+	"encoding/binary"
+	"errors"
+	"time"
 
-	"github.com/skycoin/cxo/data/cxds"
-	"github.com/skycoin/cxo/data/idxdb"
+	"github.com/skycoin/skycoin/src/cipher"
 )
 
-// A CXDS is interface of CX data store that is client
-// for CX data server or any stub package. The CXDS is
-// key-value store with references count
-type CXDS interface {
-	// Get value by key. Result is value and references count
-	Get(key cipher.SHA256) (val []byte, rc uint32, err error)
-	// GetInc is the same as Inc+Get
-	GetInc(key cipher.SHA256) (val []byte, rc uint32, err error)
-	// Set key-value pair. If value already exists the Set
-	// increments references count
-	Set(key cipher.SHA256, val []byte) (rc uint32, err error)
-	// Add value, calculating hash internally. If
-	// value already exists the Add increments references
-	// count
-	Add(val []byte) (key cipher.SHA256, rc uint32, err error)
-	// Inc increments references count
-	Inc(key cipher.SHA256) (rc uint32, err error)
-	// Dec decrements references count. The rc is zero if
-	// value has been deleted by the Dec
-	Dec(key cipher.SHA256) (rc uint32, err error)
-	// DecGet is the same as Dec but it returns value even if
-	// it has been deleted by call
-	DecGet(key cipher.SHA256) (val []byte, rc uint32, err error)
+// common errors
+var (
+	ErrNotFound       = errors.New("not found")
+	ErrStopIteration  = errors.New("stop iteration")
+	ErrFeedIsNotEmpty = errors.New("can't remove feed: feed is not empty")
+	ErrNoSuchFeed     = errors.New("no such feed")
 
-	// batch operation
+	ErrInvalidSize = errors.New("invalid size of encoded data")
+)
 
-	// MultiGet returns values by keys
-	MultiGet(keys []cipher.SHA256) (vals [][]byte, err error)
-	// MultiAdd append given values
-	MultiAdd(vals [][]byte) (err error)
-
-	// MultiInc increments
-	MultiInc(keys []cipher.SHA256) (err error)
-	// MultiDec decrements
-	MultiDec(keys []cipher.SHA256) (err error)
-
-	// Clsoe the CXDS
-	Close() (err error)
-}
-
-// A DB represents joiner of
-// IdxDB and CXDS
+// A DB represents joiner of IdxDB and CXDS
 type DB struct {
 	cxds  CXDS
-	idxdb idxdb.IdxDB
+	idxdb IdxDB
 }
 
 // IdxDB of the DB
-func (d *DB) IdxDB() idxdb.IdxDB {
+func (d *DB) IdxDB() IdxDB {
 	return d.idxdb
 }
 
@@ -64,6 +34,106 @@ func (d *DB) CXDS() CXDS {
 	return d.cxds
 }
 
+// Clsoe the DB and all underlying
+func (d *DB) Close() (err error) {
+	if err = d.cxds.Close(); err != nil {
+		d.idxdb.Close() // drop error
+	} else {
+		err = d.idxdb.Close() // use this error
+	}
+	return
+}
+
+// NewDB creates DB by given CXDS and IdxDB.
+// The arguments must not be nil
+func NewDB(cxds CXDS, idxdb IdxDB) *DB {
+	if cxds == nil {
+		panic("missing CXDS")
+	}
+	if idxdb == nil {
+		panic("missing IdxDB")
+	}
+	return &DB{cxds, idxdb}
+}
+
+// A Root represetns meta information
+// of a saved skyobject.Root
+type Root struct {
+	CreateTime int64 // created at, unix nano
+	AccessTime int64 // last access time, unix nano
+
+	Seq  uint64        // seq number of this Root
+	Prev cipher.SHA256 // previous Root or empty if seq == 0
+
+	Hash cipher.SHA256 // hash of the Root
+	Sig  cipher.Sig    // signature of the Root
+}
+
+// Validate the Root
+func (r *Root) Validate() (err error) {
+	if r.Seq == 0 {
+		if r.Prev != (cipher.SHA256{}) {
+			return errors.New("(idxdb.Root.Validate) unexpected Prev hash")
+		}
+	} else if r.Prev == (cipher.SHA256{}) {
+		return errors.New("(idxdb.Root.Validate) missing Prev hash")
+	}
+
+	if r.Hash == (cipher.SHA256{}) {
+		return errors.New("(idxdb.Root.Validate) empty Hash")
+	}
+
+	if r.Sig == (cipher.Sig{}) {
+		return errors.New("(idxdb.Root.Validate) empty Sig")
+	}
+	return
+}
+
+// Encode the Root
+func (r *Root) Encode() (p []byte) {
+
+	// the method genertes bytes equal to genrated by
+	// github.com/skycoin/skycoin/src/cipher/encoder
+	// but the Encode doesn't mess with reflection
+
+	p = make([]byte, 8+8+8+len(cipher.SHA256{})*2+len(cipher.Sig{}))
+
+	binary.LittleEndian.PutUint64(p, uint64(r.CreateTime))
+	binary.LittleEndian.PutUint64(p[8:], uint64(r.AccessTime))
+
+	binary.LittleEndian.PutUint64(p[8+8:], r.Seq)
+
+	copy(p[8+8+8:], r.Prev[:])
+	copy(p[8+8+8+len(cipher.SHA256{}):], r.Hash[:])
+	copy(p[8+8+8+len(cipher.SHA256{})*2:], r.Sig[:])
+
+	return
+}
+
+// Decode given encoded Root to this one
+func (r *Root) Decode(p []byte) (err error) {
+
+	if len(p) != 8+8+8+len(cipher.SHA256{})*2+len(cipher.Sig{}) {
+		return ErrInvalidSize
+	}
+
+	r.CreateTime = int64(binary.LittleEndian.Uint64(p))
+	r.AccessTime = int64(binary.LittleEndian.Uint64(p[8:]))
+
+	r.Seq = binary.LittleEndian.Uint64(p[8+8:])
+
+	copy(r.Prev[:], p[8+8+8:])
+	copy(r.Hash[:], p[8+8+8+len(cipher.SHA256{}):])
+	copy(r.Sig[:], p[8+8+8+len(cipher.SHA256{})*2:])
+	return
+}
+
+// UpdateAccessTime updates access time
+func (r *Root) UpdateAccessTime() {
+	r.AccessTime = time.Now().UnixNano()
+}
+
+/*
 // NewMemoryDB returns DB in memory
 func NewMemoryDB() (db *DB) {
 	db = new(DB)
@@ -85,18 +155,4 @@ func NewDriveDB(prefix string) (db *DB, err error) {
 	}
 	return
 }
-
-func (d *DB) Close() (err error) {
-	if err = d.cxds.Close(); err != nil {
-		d.idxdb.Close() // drop error
-	} else {
-		err = d.idxdb.Close() // use this error
-	}
-	return
-}
-
-// NewDB creates DB by given CXDS and IdxDB. The arguments
-// must not be nil
-func NewDB(cxds CXDS, idxdb idxdb.IdxDB) *DB {
-	return &DB{cxds, idxdb}
-}
+*/
