+++ data/idxdb/idxdb.go
@@ -1,156 +1,13 @@
 package idxdb
 
 import (
-	"encoding/binary"
 	"errors"
-	"time"
-
-	"github.com/skycoin/skycoin/src/cipher"
 )
 
 // common errors
 var (
-	ErrInvalidSize    = errors.New("invalid size of encoded obejct")
-	ErrStopIteration  = errors.New("stop iteration")
-	ErrFeedIsNotEmpty = errors.New("can't remove feed: feed is not empty")
-	ErrNoSuchFeed     = errors.New("no such feed")
-	ErrNotFound       = errors.New("not found")
+	ErrInvalidSize = errors.New("invalid size of encoded obejct")
 )
-
-// An IterateFeedsFunc represetns ...
-type IterateFeedsFunc func(cipher.PubKey) error
-
-// A Feeds represetns bucket of feeds
-type Feeds interface {
-	// Add feed. Adding a feed twice or
-	// mote times does nothing
-	Add(cipher.PubKey) error
-	// Del feed if its empty. It's impossible to
-	// delete non-empty feed. This restriction required
-	// for related obejcts. We need to decrement refs count
-	// of all related obejcts. Del never returns 'not found'
-	// error
-	Del(cipher.PubKey) error
-
-	Iterate(IterateFeedsFunc) error // iterate all feeds
-	Has(cipher.PubKey) bool         // presence check
-
-	// Roots of feed. You'll got ErrNoSuchFeed
-	// if given feed doesn't exist
-	Roots(cipher.PubKey) (Roots, error)
-}
-
-// An IterateRootsFunc represents ...
-type IterateRootsFunc func(*Root) error
-
-// A Roots represents bucket of Root objects
-type Roots interface {
-	Ascend(IterateRootsFunc) error  // iterate ascending oreder
-	Descend(IterateRootsFunc) error // iterate descending order
-
-	// Set or update Root. Method modifies orginal Root
-	// setting AccessTime and CreateTime to appropriate
-	// values. If Root already exists, then this method
-	// updates access time only
-	Set(*Root) error
-
-	// Del Root by seq number
-	Del(uint64) error
-
-	// Get Root by seq
-	Get(uint64) (*Root, error)
-
-	// Has the Roots Root with given seq?
-	Has(uint64) bool
-
-	// Len returns amount of saved Root oebjcts
-	Len() int
-}
-
-// An IdxDB repesents API of the index-DB
-type IdxDB interface {
-	Tx(func(Feeds) error) error // lookup transaction
-	Close() error               // close the DB
-
-	// TODO: stat
-}
-
-// A Root represetns meta information
-// of Root obejct
-type Root struct {
-	CreateTime int64 // created at, unix nano
-	AccessTime int64 // last access time, unix nano
-
-	Seq  uint64        // seq number of this Root
-	Prev cipher.SHA256 // previous Root or empty if seq == 0
-
-	Hash cipher.SHA256 // hash of the Root filed
-	Sig  cipher.Sig    // signature of the Hash field
-}
-
-// Validate the Root
-func (r *Root) Validate() (err error) {
-	if r.Seq == 0 {
-		if r.Prev != (cipher.SHA256{}) {
-			return errors.New("(idxdb.Root.Validate) unexpected Prev hash")
-		}
-	} else if r.Prev == (cipher.SHA256{}) {
-		return errors.New("(idxdb.Root.Validate) missing Prev hash")
-	}
-
-	if r.Hash == (cipher.SHA256{}) {
-		return errors.New("(idxdb.Root.Validate) empty Hash")
-	}
-
-	if r.Sig == (cipher.Sig{}) {
-		return errors.New("(idxdb.Root.Validate) empty Sig")
-	}
-	return
-}
-
-// Encode the Root
-func (r *Root) Encode() (p []byte) {
-
-	// the method genertes bytes equal to genrated by
-	// github.com/skycoin/skycoin/src/cipher/encoder
-	// but the Encode doesn't mess with reflection
-
-	p = make([]byte, 8+8+8+len(cipher.SHA256{})*2+len(cipher.Sig{}))
-
-	binary.LittleEndian.PutUint64(p, uint64(r.CreateTime))
-	binary.LittleEndian.PutUint64(p[8:], uint64(r.AccessTime))
-
-	binary.LittleEndian.PutUint64(p[8+8:], r.Seq)
-
-	copy(p[8+8+8:], r.Prev[:])
-	copy(p[8+8+8+len(cipher.SHA256{}):], r.Hash[:])
-	copy(p[8+8+8+len(cipher.SHA256{})*2:], r.Sig[:])
-
-	return
-}
-
-// Decode given encoded Root to this one
-func (r *Root) Decode(p []byte) (err error) {
-
-	if len(p) != 8+8+8+len(cipher.SHA256{})*2+len(cipher.Sig{}) {
-		return ErrInvalidSize
-	}
-
-	r.CreateTime = int64(binary.LittleEndian.Uint64(p))
-	r.AccessTime = int64(binary.LittleEndian.Uint64(p[8:]))
-
-	r.Seq = binary.LittleEndian.Uint64(p[8+8:])
-
-	copy(r.Prev[:], p[8+8+8:])
-	copy(r.Hash[:], p[8+8+8+len(cipher.SHA256{}):])
-	copy(r.Sig[:], p[8+8+8+len(cipher.SHA256{})*2:])
-	return
-}
-
-// UpdateAccessTime updates access time
-func (r *Root) UpdateAccessTime() {
-	r.AccessTime = time.Now().UnixNano()
-}
 
 /*
